+++ src/dht/routing.c
@@ -3706,8 +3706,8 @@ dht_compute_size_estimate_2(patricia_t *pt, const kuid_t *kuid)
 	estimate = (guint64) (retained * pow(2.0, bits));
 
 	if (GNET_PROPERTY(dht_debug)) {
-		g_debug("DHT average common prefix is %lf bits over %u node%s",
-			bits, retained, 1 == retained ? "" : "s");
+		g_debug("DHT average common prefix is %lf bits over %lu node%s",
+			(unsigned long) bits, retained, 1 == retained ? "" : "s");
 	}
 
 	return estimate;
